Veeam-Zabbix-Simple
Simple way to monitor Veeam backup in active mode with Zabbix.

It use a small powershell script with zabbix sender to push information to Zabbix server with a Zabbix discovery template.

It use the task scheduler ( a template is included in the folder )

You can watch if a job is in success, warning or failed.